housekeeping: split activation related classes into one file per class
