+++ src/ReactiveUI/RegisterableInterfaces.cs
@@ -1,17 +1,13 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reactive.Concurrency;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Input;
-using System.Reactive;
-using Splat;
 using System.Linq.Expressions;
+using System.Reactive;
+using System.Reactive.Concurrency;
+using System.Windows.Input;
+using Splat;
 
 namespace ReactiveUI
 {
@@ -317,17 +313,6 @@ namespace ReactiveUI
         /// <param name="viewModel">View model.</param>
         /// <param name="contract">Contract.</param>
         IViewFor ResolveView<T>(T viewModel, string contract = null) where T : class;
-    }
-
-    /// <summary>
-    /// Implement this interface to override how ReactiveUI determines when a
-    /// View is activated or deactivated. This is usually only used when porting
-    /// ReactiveUI to a new UI framework
-    /// </summary>
-    public interface IActivationForViewFetcher
-    {
-        int GetAffinityForView(Type view);
-        IObservable<bool> GetActivationForView(IActivatable view);
     }
 
     internal interface IPlatformOperations
